Allow passing of data attributes with tags
This allows data attributes added to li's in the markup as well as attributes
added to autocomplete entries to be added to tags.

Example use cases are when the tag name itself is not the unique id needed for
persistence or when you need to have a url for each tag that can be triggered on
click.