fix inconsistent autocomplete data entry
comma will still create new tags by default.

improved handling of quoted values. still possible for quotes to appear
within, and at end of a tag, as previously.

delimiter now recognised as tag separator. allows commas in a tag when
delimiter is not the default. allows complex, multiple character
separators, which can enable punctuation in tags.

was not possible to enter a tag if it was same as the beginning of an
autocomplete option. the input would just lock.

no tag could be entered if autocomplete was open, regardless of whether
autocomplete options were chosen.

could break data in hidden input if delimiter was enclosed in quotes.

Signed-off-by: ucavus <ucavus@github.fake>