Handling pasted values with the inputfieldwidth and acceptableCharsRegex
jquery "Paste" event handles the field width.  I also put the
acceptableCharsRegex down within the createTag function so the user can
see what they accidentally got into the box during the paste.

Note: eventually it would be a good idea to add a jquery Validate hook
for the entry box to be able to show validation messages on why certain
issues stop the creation of a tag.