Remove regular expression that was causing wrong behaviour
Tags can be created when the user selects comma, enter or tab key.
In this situations, there are some checks done, one of them is removing
starting "s" from entered text and then evaluate if the first character is a
quote("). Since the plugin needs to check only if the first character from
the text is a quote(") and not remove any other letters, I removed the replace
function call.