tag-it.js object support
Tag-it.js does not support custom js object and jquery autocomplete object. It shows always value of autocomplete object. I think that is wrong. It shows label of object and it is unnecessary to support default values with input and li elements. An option parameter with name items can be used as default values and custom object can be used owing to this.